This update is primarily to clean up the refresh phase.
Build
  -added temp_new_refresh
  -rearranged link order for Refresh object in Block
Include
  -rearranged include order for Refresh object in Block
Method
  -added refresh_ object for method refresh
  -added refresh_ initialization to Enzo Methods
  -added FieldDescr argument to EnzoMethodHeat constructor (for refresh_
   initialization
Block
  -added refresh_ object for current Refresh operation
  -removed refresh_call_, refresh_sync_, index_refresh_
Initialize
  -removed initialize_refresh()
Refresh
  -updated Refresh synchronization to count neighbors (using current
   Block::refresh_ Sync), not use "neighbor" sync on p_refresh_exit()
  -added count parameter to refresh_load_face_ and refresh_store_face_()
  -removed name from Refresh object
  -added Block::sync_, is_active_, callback_;
  -renamed field_ghost_depth_ -> ghost_depth_

Balance
  -added quiescence call before AtSync()
Charm
  -added p_stopping_balance() entry method for quiescence call before AtSync()
Error
  -fixed comparing char arrays using ==
Problem
  -moved ScalarExpr::evalute() template from hpp to cpp
  -moved Value::evaluate() template from hpp to cpp
Gravity
  -added missing p_mg0_*() entry methods to enzo.ci
Testing
  -renamed redundant "Mesh" section in test/index.php as "Adapt"

+++ src/Cello/memory_Memory.cpp
@@ -256,7 +256,7 @@ int Memory::num_delete ( std::string group_name ) throw ()
 void Memory::print () throw ()
 {
 #ifdef CONFIG_USE_MEMORY
-  for (int i=0; i< group_name_.size(); i++) {
+  for (size_t i=0; i< group_name_.size(); i++) {
     Monitor * monitor = Monitor::instance();
     if (i == 0 || group_name_[i] != "") {
       monitor->print ("Memory","Group %s",i ? group_name_[i].c_str(): "Total");
@@ -278,7 +278,7 @@ void Memory::reset() throw()
 #ifdef CONFIG_USE_MEMORY
   index_group_ = 0;
 
-  for (int i=0; i<bytes_curr_.size(); i++) {
+  for (size_t i=0; i<bytes_curr_.size(); i++) {
     bytes_curr_    [i] = 0;
     bytes_high_    [i] = 0;
     bytes_highest_ [i] = 0;
@@ -294,7 +294,7 @@ void Memory::reset_high() throw()
 {
 #ifdef CONFIG_USE_MEMORY
   TRACE("reset_high");
-  for (int i=0; i<bytes_high_.size(); i++) {
+  for (size_t i=0; i<bytes_high_.size(); i++) {
     bytes_high_ [i] = bytes_curr_[i];
   }
 #endif
